pyproject.toml
[tool.poetry]
name = "fidmaa"
version = "0.1.0"
description = "What's FIDMAA?"
authors = ["Michał Pasternak <michal.dtz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
PySide6 = "^6.4.1"
Pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
